requirements.txt
streamlit>=1.35.0
google-genai>=0.7.0
pdfplumber>=0.11.0
python-docx>=1.1.0
wordcloud>=1.9.3
matplotlib>=3.8.0
python-dotenv>=1.0.0